Normalize volume using EBU R 128 integrated loudness
EBU R 128 is an industry-standard recommendation that defines an
objective algorithm for quantifying and normalizing loudness. This is
the same standard that platforms like YouTube, Spotify, and Apple Music
supposedly use.

As opposed to a naive average power calculation, the EBU R 128 algorithm
weights frequencies according to the human ear's sensitivity to them,
and also detects and ignores periods of silence within a waveform. The
resulting "loudness" is measured in LUFS, which are the same size as a
decibel. The recommended normalized loudness by EBU R 128 is -23 LUFS.