Fix shuffle time bound calculation
Currently bounds on repeat count do not factor in the repeat offset
(loop start), so the durations corresponding to a given number of
repeats is underestimated. Fix this bug.

Additionally, shift calculated bounds by shuffle variance, in order to
ensure that the bounds will be fully respected once shuffle variance has
been applied.